+++ src/components/quiz/preparation/components/QuizSettings.module.css
@@ -1,4 +1,4 @@
-@import '../../../config/variables.css';
+@import '../../../../config/variables.css';
 
 .quizSettings {
   background: var(--color-surface-1);
