+++ src/components/quiz/preparation/components/SelectedCharactersList.jsx
