+++ src/components/quiz/preparation/hiragana-table/CategoryInfo.jsx
