+++ src/components/quiz/preparation/hiragana-table/CategoryInfo.module.css
@@ -1,4 +1,4 @@
-@import '../../../config/variables.css';
+@import '../../../../config/variables.css';
 
 /* 카테고리 설명 */
 .categoryDescription {
