+++ src/components/quiz/preparation/hiragana-table/CategoryTabs.module.css
@@ -1,4 +1,4 @@
-@import '../../../config/variables.css';
+@import '../../../../config/variables.css';
 
 /* 탭 컨테이너 */
 .tabsContainer {
