+++ src/components/quiz/preparation/hiragana-table/ExtendedHiraganaTable.module.css
@@ -1,4 +1,4 @@
-@import '../../../config/variables.css';
+@import '../../../../config/variables.css';
 
 .extendedHiraganaTable {
   width: 100%;
