+++ src/components/quiz/preparation/hiragana-table/extendedHiraganaData.js
