--- src/data/hiraganaData.js
@@ -1,114 +1,0 @@
-export const hiraganaData = [
-  {
-    row: 'あ',
-    characters: [
-      { hiragana: 'あ', romaji: 'a', image: 'https://via.placeholder.com/150/FF0000/FFFFFF?text=A', description: 'This is a placeholder description for あ.' },
-      { hiragana: 'い', romaji: 'i', image: 'https://via.placeholder.com/150/00FF00/FFFFFF?text=I', description: 'This is a placeholder description for い.' },
-      { hiragana: 'う', romaji: 'u', image: 'https://via.placeholder.com/150/0000FF/FFFFFF?text=U', description: 'This is a placeholder description for う.' },
-      { hiragana: 'え', romaji: 'e', image: 'https://via.placeholder.com/150/FFFF00/000000?text=E', description: 'This is a placeholder description for え.' },
-      { hiragana: 'お', romaji: 'o', image: 'https://via.placeholder.com/150/FF00FF/FFFFFF?text=O', description: 'This is a placeholder description for お.' }
-    ]
-  },
-  {
-    row: 'か',
-    characters: [
-      { hiragana: 'か', romaji: 'ka', image: 'https://via.placeholder.com/150/00FFFF/000000?text=KA', description: 'This is a placeholder description for か.' },
-      { hiragana: 'き', romaji: 'ki', image: 'https://via.placeholder.com/150/800000/FFFFFF?text=KI', description: 'This is a placeholder description for き.' },
-      { hiragana: 'く', romaji: 'ku', image: 'https://via.placeholder.com/150/008000/FFFFFF?text=KU', description: 'This is a placeholder description for く.' },
-      { hiragana: 'け', romaji: 'ke', image: 'https://via.placeholder.com/150/000080/FFFFFF?text=KE', description: 'This is a placeholder description for け.' },
-      { hiragana: 'こ', romaji: 'ko', image: 'https://via.placeholder.com/150/808000/FFFFFF?text=KO', description: 'This is a placeholder description for こ.' }
-    ]
-  },
-  {
-    row: 'さ',
-    characters: [
-      { hiragana: 'さ', romaji: 'sa', image: '', description: '' },
-      { hiragana: 'し', romaji: 'shi', image: '', description: '' },
-      { hiragana: 'す', romaji: 'su', image: '', description: '' },
-      { hiragana: 'せ', romaji: 'se', image: '', description: '' },
-      { hiragana: 'そ', romaji: 'so', image: '', description: '' }
-    ]
-  },
-  {
-    row: 'た',
-    characters: [
-      { hiragana: 'た', romaji: 'ta', image: '', description: '' },
-      { hiragana: 'ち', romaji: 'chi', image: '', description: '' },
-      { hiragana: 'つ', romaji: 'tsu', image: '', description: '' },
-      { hiragana: 'て', romaji: 'te', image: '', description: '' },
-      { hiragana: 'と', romaji: 'to', image: '', description: '' }
-    ]
-  },
-  {
-    row: 'な',
-    characters: [
-      { hiragana: 'な', romaji: 'na', image: '', description: '' },
-      { hiragana: 'に', romaji: 'ni', image: '', description: '' },
-      { hiragana: 'ぬ', romaji: 'nu', image: '', description: '' },
-      { hiragana: 'ね', romaji: 'ne', image: '', description: '' },
-      { hiragana: 'の', romaji: 'no', image: '', description: '' }
-    ]
-  },
-  {
-    row: 'は',
-    characters: [
-      { hiragana: 'は', romaji: 'ha', image: '', description: '' },
-      { hiragana: 'ひ', romaji: 'hi', image: '', description: '' },
-      { hiragana: 'ふ', romaji: 'fu', image: '', description: '' },
-      { hiragana: 'へ', romaji: 'he', image: '', description: '' },
-      { hiragana: 'ほ', romaji: 'ho', image: '', description: '' }
-    ]
-  },
-  {
-    row: 'ま',
-    characters: [
-      { hiragana: 'ま', romaji: 'ma', image: '', description: '' },
-      { hiragana: 'み', romaji: 'mi', image: '', description: '' },
-      { hiragana: 'む', romaji: 'mu', image: '', description: '' },
-      { hiragana: 'め', romaji: 'me', image: '', description: '' },
-      { hiragana: 'も', romaji: 'mo', image: '', description: '' }
-    ]
-  },
-  {
-    row: 'や',
-    characters: [
-      { hiragana: 'や', romaji: 'ya', image: '', description: '' },
-      null,
-      { hiragana: 'ゆ', romaji: 'yu', image: '', description: '' },
-      null,
-      { hiragana: 'よ', romaji: 'yo', image: '', description: '' }
-    ]
-  },
-  {
-    row: 'ら',
-    characters: [
-      { hiragana: 'ら', romaji: 'ra', image: '', description: '' },
-      { hiragana: 'り', romaji: 'ri', image: '', description: '' },
-      { hiragana: 'る', romaji: 'ru', image: '', description: '' },
-      { hiragana: 'れ', romaji: 're', image: '', description: '' },
-      { hiragana: 'ろ', romaji: 'ro', image: '', description: '' }
-    ]
-  },
-  {
-    row: 'わ',
-    characters: [
-      { hiragana: 'わ', romaji: 'wa', image: '', description: '' },
-      null,
-      null,
-      null,
-      { hiragana: 'を', romaji: 'wo', image: '', description: '' }
-    ]
-  },
-  {
-    row: 'ん',
-    characters: [
-      null,
-      null,
-      { hiragana: 'ん', romaji: 'n', image: '', description: '' },
-      null,
-      null
-    ]
-  }
-];
-
-export const columns = ['あ', 'い', 'う', 'え', 'お'];
